Peerasanti 1422

+++ InitialProject/SourceCode/database/migrations/2022_03_15_141518_create_programs_table.php
@@ -13,15 +13,17 @@ class CreateProgramsTable extends Migration
      */
     public function up()
     {
-        Schema::create('programs', function (Blueprint $table) {
-            $table->id();
-            $table->string('program_name_th');
-            $table->string('program_name_en');
+        if (!Schema::hasTable('programs')) {
+            Schema::create('programs', function (Blueprint $table) {
+                $table->id();
+                $table->string('program_name_th');
+                $table->string('program_name_en');
 
-            $table->unsignedBigInteger('degree_id');
-            $table->foreign('degree_id')->references('id')->on('degrees')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->unsignedBigInteger('degree_id');
+                $table->foreign('degree_id')->references('id')->on('degrees')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
